requirements.txt
torch
torchvision
transformers
gradio
Pillow
numpy
datasets
